Fix compilation warnings on Win64
Fix compilation warnings on Win64, as shown below, by using proper long types
```
  gcompletion.c(47): warning C4028: formal parameter 3 different from declaration [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(105): warning C4133: 'function': incompatible types - from 'gsize *' to 'size_t *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gconvert.c(460): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gdataset.c(552): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  giochannel.c(1238): warning C4133: 'function': incompatible types - from 'size_t *' to 'gsize *' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  gmem.c(662): warning C4311: 'type cast': pointer truncation from 'gpointer' to 'guint' [C:\Users\bellenot\build\x64\release\graf2d\win32gdk\glib.vcxproj]
  ...
```

+++ graf2d/win32gdk/gdk/src/glib/giochannel.c
@@ -1722,7 +1722,7 @@ g_io_channel_read_chars (GIOChannel	*channel,
 
   if (!channel->use_buffer)
     {
-      gint tmp_bytes;
+      gsize tmp_bytes;
       
       g_assert (!channel->read_buf || channel->read_buf->len == 0);
 
@@ -1914,7 +1914,7 @@ g_io_channel_write_chars (GIOChannel	*channel,
 
   if (!channel->use_buffer)
     {
-      gint tmp_bytes;
+      gsize tmp_bytes;
       
       g_assert (!channel->write_buf || channel->write_buf->len == 0);
       g_assert (channel->partial_write_buf[0] == '\0');
